FIX WARNING

+++ Tests/TestNetworkClient.cpp
@@ -13,7 +13,7 @@ int main()
     std::vector<std::unique_ptr<CClient>> Clients;
     std::vector<std::future<std::shared_ptr<CConnection>>> ClientFutureConnections;
     std::vector<std::shared_ptr<CConnection>> ClientConnections;
-    for (size_t i = 0; i < TestClientCounter; i++)
+    for (int32_t i = 0; i < TestClientCounter; i++)
     {
         Clients.emplace_back(std::make_unique<CClient>(1));
         ClientFutureConnections.push_back(Clients[i]->ConnectToServer("127.0.0.1", TEST_PORT));
